first repair after repo loss
