+++ include/web_interface.h
@@ -1,63 +1,84 @@
 #ifndef WEB_INTERFACE_H
 #define WEB_INTERFACE_H
 
-#include <WebServer.h>
-#include <ArduinoJson.h>
-#include <ESPmDNS.h>
+#include <Arduino.h>
 #include "thermostat_state.h"
 #include "config_manager.h"
 #include "sensor_interface.h"
 #include "knx_interface.h"
 #include "mqtt_interface.h"
 #include "pid_controller.h"
-#include <LittleFS.h>
+#include "protocol_manager.h"
+
+// Command and source definitions for protocol manager
+#define CMD_SET_TEMPERATURE 1
+#define SOURCE_WEB_API 1
+
+#ifdef ESP32
+  #include <WebServer.h>
+  #include <ESPmDNS.h>
+  #include <LITTLEFS.h>
+  #define WebServerClass WebServer
+  #define LittleFS LITTLEFS
+#elif defined(ESP8266)
+  #include <ESP8266WebServer.h>
+  #include <ESP8266mDNS.h>
+  #include <LittleFS.h>
+  #define WebServerClass ESP8266WebServer
+#endif
 
 class WebInterface {
 public:
-    WebInterface();
-    bool begin(ThermostatState* state, 
-               ConfigManager* config, 
-               SensorInterface* sensor,
-               KNXInterface* knx,
-               MQTTInterface* mqtt,
-               PIDController* pid);
-    void handle();
+  // Constructor
+  WebInterface();
+  
+  // Initialize web server
+  bool begin(ThermostatState* state, 
+             ConfigManager* config, 
+             SensorInterface* sensor,
+             KNXInterface* knx,
+             MQTTInterface* mqtt = nullptr,
+             PIDController* pid = nullptr, 
+             ProtocolManager* protocol = nullptr);
+  
+  // Process HTTP requests (call in loop)
+  void handle();
 
 private:
-    // References to components
-    ThermostatState* thermostatState;
-    ConfigManager* configManager;
-    SensorInterface* sensorInterface;
-    KNXInterface* knxInterface;
-    MQTTInterface* mqttInterface;
-    PIDController* pidController;
-    
-    // WebServer instance
-    WebServer server;
-
-    // HTTP request handlers
-    void handleRoot();
-    void handleGetStatus();
-    void handleSetpoint();
-    void handleSaveConfig();
-    void handleReboot();
-    void handleFactoryReset();
-    void handleNotFound();
+  // Web server
+  WebServerClass server;
   
-    // Setup MDNS
-    void setupMDNS();
-    
-    // Serve static files from LittleFS
-    bool handleFileRead(String path);
-    
-    // Authentication helpers
-    bool isAuthenticated();
-    void requestAuthentication();
-    
-    // Rate limiting
-    unsigned long lastRequestTime;
-    int requestCount;
-    static const int MAX_REQUESTS_PER_MINUTE = 60;
+  // References to components
+  ThermostatState* thermostatState;
+  ConfigManager* configManager;
+  SensorInterface* sensorInterface;
+  KNXInterface* knxInterface;
+  MQTTInterface* mqttInterface;
+  PIDController* pidController;
+  ProtocolManager* protocolManager;
+  
+  // HTTP request handlers
+  void handleRoot();
+  void handleSave();
+  void handleSaveConfig();
+  void handleSetpoint();
+  void handleReboot();
+  void handleFactoryReset();
+  void handleNotFound();
+  void handleGetStatus();
+  
+  // Helper to generate HTML
+  String generateHtml();
+  
+  // Setup MDNS
+  void setupMDNS();
+  
+  // File handling
+  bool handleFileRead(String path);
+  
+  // Authentication
+  bool isAuthenticated();
+  void requestAuthentication();
 };
 
 #endif // WEB_INTERFACE_H
