+++ platformio.ini
@@ -3,6 +3,7 @@ default_envs = esp32
 src_dir = src
 include_dir = include
 
+
 [env:esp32]
 platform = espressif32
 board = nodemcu-32s
@@ -20,6 +21,6 @@ lib_deps =
     https://github.com/tzapu/WiFiManager.git#v2.0.15-rc.1
 	bblanchon/ArduinoJson @ ^6.21.3
     knolleary/PubSubClient @ ^2.8.0
-
+    lorol/LittleFS_esp32 @ ^1.0.6
 upload_speed = 115200
 monitor_filters = esp32_exception_decoder
