+++ src/communication/protocol_manager.cpp
@@ -1,129 +1,144 @@
+#include "thermostat_state.h"
 #include "protocol_manager.h"
 #include "knx_interface.h"
 #include "mqtt_interface.h"
 
-ProtocolManager::ProtocolManager() : 
-  thermostatState(nullptr),
+ProtocolManager::ProtocolManager() :
   knxInterface(nullptr),
   mqttInterface(nullptr),
-  lastCommandSource(SOURCE_LOCAL),
-  lastCommandTime(0) {
+  thermostatState(nullptr),
+  lastSendTime(0),
+  sendInterval(10000) { // Default 10 seconds
 }
 
 void ProtocolManager::begin(ThermostatState* state) {
   thermostatState = state;
 }
 
-void ProtocolManager::registerProtocols(KNXInterface* knx, MQTTInterface* mqtt) {
+void ProtocolManager::registerProtocols(KNXInterface* knx) {
   knxInterface = knx;
+  
+  // Register callbacks if state is available
+  if (thermostatState && knxInterface) {
+    knxInterface->registerCallbacks(thermostatState);
+  }
+}
+
+void ProtocolManager::registerMQTT(MQTTInterface* mqtt) {
   mqttInterface = mqtt;
-}
-
-void ProtocolManager::handleIncomingCommand(CommandSource source, CommandType cmd, float value) {
-  // Skip processing if thermostat state not available
-  if (!thermostatState) {
-    Serial.println("Cannot process command: thermostat state not available");
-    return;
-  }
-
-  // Prevent command loops by implementing a cooldown
-  unsigned long currentTime = millis();
-  if (currentTime - lastCommandTime < COMMAND_COOLDOWN) {
-    // Only allow if new command has higher priority
-    if (!hasPriority(source, lastCommandSource)) {
-      Serial.printf("Command rejected: %d cooldown active, source %d doesn't have priority over %d\n", 
-                  COMMAND_COOLDOWN, source, lastCommandSource);
-      return;
-    }
-  }
   
-  // Update last command details
-  lastCommandSource = source;
-  lastCommandTime = currentTime;
-  
-  // Process command based on type
-  switch (cmd) {
-    case CMD_SET_TEMPERATURE:
-      Serial.printf("Setting target temperature to %.2f°C from source %d\n", value, source);
-      thermostatState->setTargetTemperature(value);
-      break;
-      
-    case CMD_SET_MODE:
-      if (value >= MODE_OFF && value <= MODE_ANTIFREEZE) {
-        Serial.printf("Setting mode to %d from source %d\n", (int)value, source);
-        thermostatState->setMode(static_cast<ThermostatMode>(static_cast<int>(value)));
-      } else {
-        Serial.printf("Invalid mode value: %d\n", (int)value);
-      }
-      break;
-      
-    case CMD_SET_VALVE_POSITION:
-      // Only certain sources should be able to directly set valve position
-      if (source == SOURCE_LOCAL || source == SOURCE_WEB_API) {
-        Serial.printf("Setting valve position to %.2f%% from source %d\n", value, source);
-        thermostatState->setValvePosition(value);
-      } else {
-        Serial.printf("Valve position command from source %d rejected (unauthorized source)\n", source);
-      }
-      break;
-      
-    default:
-      Serial.printf("Unknown command type: %d\n", cmd);
-      break;
-  }
-  
-  // Broadcast state to all interfaces after change
-  broadcastState(thermostatState);
-}
-
-void ProtocolManager::broadcastState(ThermostatState* state) {
-  // Skip if state not available
-  if (!state) return;
-  
-  // Broadcast current state to all protocols
-  if (knxInterface && state) {
-    knxInterface->sendTemperature(state->currentTemperature);
-    knxInterface->sendSetpoint(state->targetTemperature);
-    knxInterface->sendValvePosition(state->valvePosition);
-    knxInterface->sendMode(state->operatingMode);
-  }
-  
-  if (mqttInterface && state) {
-    mqttInterface->publishTemperature(state->currentTemperature);
-    mqttInterface->publishHumidity(state->currentHumidity);
-    mqttInterface->publishPressure(state->currentPressure);
-    mqttInterface->publishSetpoint(state->targetTemperature);
-    mqttInterface->publishValvePosition(state->valvePosition);
-    mqttInterface->publishMode(state->operatingMode);
-    mqttInterface->publishHeatingStatus(state->heatingActive);
-  }
+  // No special setup needed for MQTT, it registers its own callbacks
 }
 
 void ProtocolManager::loop() {
-  // Process protocol loops
-  if (knxInterface) {
-    knxInterface->loop();
-  }
-  
-  if (mqttInterface) {
-    mqttInterface->loop();
-  }
-  
-  // Periodically broadcast state (heartbeat)
-  static unsigned long lastBroadcastTime = 0;
   unsigned long currentTime = millis();
   
-  // Every 60 seconds, broadcast state as heartbeat
-  if (currentTime - lastBroadcastTime > 60000) {
-    lastBroadcastTime = currentTime;
-    
+  // Check if it's time to send updates
+  if (currentTime - lastSendTime >= sendInterval) {
+    // Send current values to protocols
     if (thermostatState) {
-      broadcastState(thermostatState);
+      sendTemperature(thermostatState->currentTemperature);
+      sendSetpoint(thermostatState->targetTemperature);
+      sendValvePosition(thermostatState->valvePosition);
+      sendMode(thermostatState->operatingMode);
     }
+    
+    lastSendTime = currentTime;
   }
 }
 
-bool ProtocolManager::hasPriority(CommandSource a, CommandSource b) {
-  // Priority order: LOCAL > KNX > MQTT > WEB_API > SCHEDULE
-  return a < b;
+bool ProtocolManager::handleIncomingCommand(int source, int command, float value) {
+  if (!thermostatState) {
+    return false;
+  }
+  
+  bool success = false;
+  
+  switch (command) {
+    case CMD_SET_TEMPERATURE:
+      thermostatState->setTargetTemperature(value);
+      // Propagate to other protocols (except the source)
+      if (source != SOURCE_KNX && knxInterface) {
+        knxInterface->sendSetpoint(value);
+      }
+      if (source != SOURCE_MQTT && mqttInterface) {
+        mqttInterface->publishSetpoint(value);
+      }
+      success = true;
+      break;
+      
+    case CMD_SET_MODE:
+      thermostatState->setMode(static_cast<ThermostatMode>(value));
+      // Propagate to other protocols
+      if (source != SOURCE_KNX && knxInterface) {
+        knxInterface->sendMode(static_cast<ThermostatMode>(value));
+      }
+      if (source != SOURCE_MQTT && mqttInterface) {
+        mqttInterface->publishMode(static_cast<ThermostatMode>(value));
+      }
+      success = true;
+      break;
+      
+    case CMD_SET_VALVE:
+      thermostatState->setValvePosition(value);
+      // Propagate to other protocols
+      if (source != SOURCE_KNX && knxInterface) {
+        knxInterface->sendValvePosition(value);
+      }
+      if (source != SOURCE_MQTT && mqttInterface) {
+        mqttInterface->publishValvePosition(value);
+      }
+      success = true;
+      break;
+  }
+  
+  return success;
+}
+
+void ProtocolManager::sendTemperature(float temperature) {
+  // Send to KNX
+  if (knxInterface) {
+    knxInterface->sendTemperature(temperature);
+  }
+  
+  // Send to MQTT
+  if (mqttInterface) {
+    mqttInterface->publishTemperature(temperature);
+  }
+}
+
+void ProtocolManager::sendSetpoint(float setpoint) {
+  // Send to KNX
+  if (knxInterface) {
+    knxInterface->sendSetpoint(setpoint);
+  }
+  
+  // Send to MQTT
+  if (mqttInterface) {
+    mqttInterface->publishSetpoint(setpoint);
+  }
+}
+
+void ProtocolManager::sendValvePosition(float position) {
+  // Send to KNX
+  if (knxInterface) {
+    knxInterface->sendValvePosition(position);
+  }
+  
+  // Send to MQTT
+  if (mqttInterface) {
+    mqttInterface->publishValvePosition(position);
+  }
+}
+
+void ProtocolManager::sendMode(ThermostatMode mode) {
+  // Send to KNX
+  if (knxInterface) {
+    knxInterface->sendMode(mode);
+  }
+  
+  // Send to MQTT
+  if (mqttInterface) {
+    mqttInterface->publishMode(mode);
+  }
 }
