+++ src/config/config_manager.cpp
@@ -1,3 +1,12 @@
+// Basic includes
+#include <Arduino.h>
+
+// Include the ThermostatState first since other components depend on it
+#include "thermostat_state.h"
+
+// Then include your component headers
+#include "knx_interface.h"
+#include "config_manager.h"
 #include "config_manager.h"
 #include <WiFiManager.h>
 #include <LittleFS.h>
